requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
djangorestframework-simplejwt>=5.3.1
djoser>=2.2.3,<2.3
mysqlclient>=2.1.0
psycopg2-binary>=2.9.9,<3.0
channels>=3.0,<4.0
channels_redis>=3.2.0,<4.0
Pillow>=8.2.0,<8.3.0
django-cors-headers>=4.3.1,<4.4
python-dotenv